Theory implementation

+++ src/SOPSCore/mod.rs
@@ -2,7 +2,7 @@ pub mod bridging;
 
 #[derive(Eq, PartialEq, Hash)]
 struct Particle {
-    x: u8,
-    y: u8,
+    x: u16,
+    y: u16,
     onland: bool,
 }
